# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LeadChain API",
        "description": "API REST de LeadChain para la gestión comercial de clientes, edificios y visitas. Autenticación mediante JWT (Bearer token). Incluye soporte para coordenadas geográficas PostGIS.",
        "contact": {
            "name": "Alberto Romero Pino",
            "email": "albertoromeropino2004@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/zonas": {
            "get": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Listar todas las zonas",
                "operationId": "cd544c2ac31c5118f9b1da4dba211784",
                "responses": {
                    "200": {
                        "description": "Listado de zonas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ZonaResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Crear una nueva zona",
                "operationId": "bd20583565934936515a593b934e3570",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ZonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zona creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZonaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear la zona"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zonas/pagina/datos": {
            "get": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Obtener datos optimizados para la página de zonas",
                "description": "Retorna todas las zonas con sus edificios y clientes en una única consulta, optimizado para la visualización en la página",
                "operationId": "86873a348bd8753666b557b77959a656",
                "responses": {
                    "200": {
                        "description": "Datos de zonas para la página",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ZonaPageResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zonas/{zona}": {
            "put": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Actualizar una zona completa",
                "operationId": "4e167d4308e586f4bca40aa2dda1a089",
                "parameters": [
                    {
                        "name": "zona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ZonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zona actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZonaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Eliminar una zona",
                "operationId": "b07d2639673abbcd9f6e47965d668f81",
                "parameters": [
                    {
                        "name": "zona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zona eliminada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Actualizar parcialmente una zona",
                "operationId": "e42696dbb1815f3b3ef3b0c37e2b7d6a",
                "parameters": [
                    {
                        "name": "zona",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zona actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ZonaResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/zonas/mapa": {
            "get": {
                "tags": [
                    "Zonas"
                ],
                "summary": "Obtener zonas, edificios y clientes para vista de mapa (una sola consulta optimizada)",
                "description": "Get consolidated data for map view\nReturns: zonas with nested edificios and clientes in a single optimized query",
                "operationId": "67b00a60288ec22872b196dad4d3fba6",
                "responses": {
                    "200": {
                        "description": "Zonas con edificios y clientes anidados. Optimizado en una única query.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "zonas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MapaZonasResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "GeoPoint": {
                "required": [
                    "lat",
                    "lng"
                ],
                "properties": {
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "example": 37.9
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "example": -4.8
                    }
                },
                "type": "object"
            },
            "ZonaInput": {
                "required": [
                    "nombre",
                    "area"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        },
                        "minItems": 4
                    }
                },
                "type": "object"
            },
...
            "ZonaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "MapaZonasResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        }
                    },
                    "edificios": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "direccion_completa": {
                                    "type": "string",
                                    "example": "Calle Gran Capitan 10"
                                },
                                "tipo": {
                                    "type": "string",
                                    "example": "residencial"
                                },
                                "id_zona": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "ubicacion": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/GeoPoint"
                                        }
                                    ],
                                    "nullable": true
                                },
                                "clientes": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Antonio"
                                            },
                                            "apellidos": {
                                                "type": "string",
                                                "example": "Perez Garcia"
                                            },
                                            "telefono": {
                                                "type": "string",
                                                "example": "612345678"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "antonio@ejemplo.com"
                                            },
                                            "planta": {
                                                "type": "integer",
                                                "example": 3,
                                                "nullable": true
                                            },
                                            "puerta": {
                                                "type": "integer",
                                                "example": 5,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ZonaPageResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        }
                    },
                    "edificios": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}